requirements.txt
# Editable install with no version control (2d-moba==0.1)
-e d:\2d moba
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.7.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
auto-py-to-exe==2.45.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bottle==0.13.2
bottle-websocket==0.2.9
Brotli==1.1.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==43.0.3
customtkinter==5.2.2
cx_Freeze==6.15.11
cx_Logging==3.2.1
darkdetect==0.8.0
decorator==4.4.2
dnspython==2.7.0
Eel==0.18.1
email_validator==2.2.0
fastapi==0.115.6
fastapi-users==14.0.0
fastapi-users-db-sqlalchemy==6.0.1
firebase-admin==6.6.0
future==1.0.0
gcloud==0.18.3
gevent==24.11.1
gevent-websocket==0.10.1
glcontext==3.0.0
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jax==0.4.38
jaxlib==0.4.38
jsons==1.6.3
jwcrypto==1.5.6
keyboard==0.13.5
keyring==25.5.0
lief==0.15.1
lxml==5.3.0
makefun==1.15.6
ml_dtypes==0.5.0
moderngl==5.12.0
moderngl-window==2.4.6
more-itertools==10.5.0
mouse==0.7.1
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==1.1.0
multipledispatch==1.0.0
mutagen==1.47.0
Nuitka==2.5.6
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.10.0.84
opt_einsum==3.4.0
ordered-set==4.1.0
packaging==24.2
pefile==2023.2.7
Pillow==10.0.0
proglog==0.1.10
proto-plus==1.25.0
protobuf==5.28.3
pwdlib==0.2.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pygame==2.5.2
pygame-ce==2.5.2
pygame_gui==0.6.12
PyGetWindow==0.0.9
pyglet==2.0.18
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2024.10
PyJWT==2.9.0
PyMsgBox==1.0.9
pynput==1.7.6
PyOpenGL==3.1.7
pyparsing==3.2.0
pyperclip==1.9.0
PyQt6==6.7.1
PyQt6-Qt6==6.7.3
PyQt6-WebEngine-Qt6==6.7.3
PyQt6-WebEngineSubwheel-Qt6==6.7.3
PyQt6_sip==13.8.0
Pyrebase4==4.8.0
PyRect==0.2.0
pyrr==0.10.3
PyScreeze==1.0.1
python-i18n==0.3.9
python-jwt==4.1.0
python-multipart==0.0.17
pytweening==1.2.0
pywin32==306
pywin32-ctypes==0.2.3
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
scipy==1.14.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.23
starlette==0.41.3
tkinterdnd2==0.4.2
tqdm==4.66.6
typing_extensions==4.12.2
typish==1.9.3
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
websockets==13.1
wheel==0.45.1
yt-dlp==2024.10.7
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
